fix: remove pointless caching for anything that is not a static SvelteKit asset
The intention was that every GET the app sends via fetch would be cached, that includes SvelteKit routes, several Firebase calls, Mapbox tiles, ... but to prevent stale cache problems, these caches would *only be used* in case of a fetch error on assumed network loss.

The problem is that they're useless for now when the network is down. Nothing loads. The app depends largely on Firebase to be able to initialize and operate, and Firebase presumably uses a WebSocket for transportation (https://stackoverflow.com/questions/49156783/transport-protocols-firebase-realtime-database-uses) which can't be cached by the SW (https://stackoverflow.com/a/39602235/4973029)