fix: make country selection a full <select> element with localized country names
Instead of the text input and + datalist approach used so far. That approach is only recommended when the datalist values are _suggestions_, not _discrete options_, as is our case. We got several reports of people being confused that their typed country name resulted in an error ("pick a value from the list"). The datalist narrows down while typing in many browsers (good for UX), but doesn't show values when you typed a country name in a different language.

Using <select> prevents these errors. To keep the UX smooth enough for most people, this also adds prefilled country guessing based on the Accept-Language header.

Also:
- shared email/password between sign-in and register pages, in case people mistakenly start registering on the sign-in page
- includes several misc behavior and style fixes for the sign-in/register flow