fix: don't reset full garden manage form after one address field edit, support house numbers in reverse geocoding (1), properly handle existing gardens in /garden/add and /garden/manage (2)
(1): House numbers were not returned from reverse geocoding, even if they were available. This lead to the "Indonesia bug": people were fine-tuning (filling in) the house number result because it was empty. When they did that, nothing happened (because the house number field was still focused). But after they clicked the "confirm button", it 1) defocussed the field, 2) reset the form because of an address edit after reverse geocoding, 2) did a geolocation lookup solely using the house number => this confirmed the unpredictable geolocation result.

(2): these pages were not waiting on the garden being loaded.
- /manage always notified "please add your garden first" because of this.
- /add didn't redirect to /manage when a garden existed