fix: work around bug on some iPhones where the html file <input> `accept` attribute is broken.
In particular, the bug does not allow .gpx files to be selected, even though the extension is mentioned in the accept attribute.

As a workaround, this does not specify (= allows) any file type on *all* iOS devices. It leads to a generic file upload dialog with media selection option everywhere (e.g. on my 11 Pro), where on iPhones where it used to work, it would jump straight to files because it recognized that photo/video type files were not accepted. It is a slightly degraded UX then for the benefit of all.