fix: Firebase Function linting
We updated the root repo to yarn@4.9.0 via corepack a while ago. In the the GH Actions pipelines corepack is enabled, but not in the Firebase Function deploy pipeline. The "yarn" command ran in the root dir which now did not have a checked-in release anymore. Using the --cwd did not make it pick up api/.yarn/releases as intended. A simple `cd` before did however (yarn v1 probably only looks for ./yarn/releases)