feat: immediately show a newly added photo after garden creation/update; remove copy informing about a delayed display
This uses optimistic UI, leveraging the image that was already loaded in the browser. This may cause issues if the the user immediately refreshes the browser after the garden creation/update, but it will cover the missing image in most cases.

The copy was also misleading and caused wrong expectations, since there was no image loader that would show the image once it became available (not even after a minute).