test: add e2e test improvements
- refactor e2e tests to more granularly and meaningfully target a Stripe demo env and demo project (vs. "local"/"staging")
- temporarily ignore unreliable Gmail email checking code (only mailpit is active now)
- fix Stripe sandbox payment automation & support the fr locale
- add support to directly assert or modify data with the Firebase admin & Stripe APIs from Playwright test code (used in one new assertion for now)
- fix a StraightToMember load state waiting issue

Note: this adds firebase-admin and stripe to th dev dependencies of the root SvelteKit project, and duplicates firebase-admin/Stripe API loading code in ES modules (in /api they are in CommonJS) which is kind of awkward. Perhaps the Playwright Node.js project could be separated, or, if the /api backend is refactored to ES modules, they could share code.